refactor: remove logic for main entry point
