+++ dundie/cli.py
@@ -1,0 +1,25 @@
+""" Command Line Interface for Dundie Mifflin Rewards """
+
+import argparse
+from dundie.core import load
+
+
+def main():
+    parser = argparse.ArgumentParser(
+        description="Dundie Mifflin Rewards CLI",
+        epilog="Enjoy and use with caution",
+    )
+
+    parser.add_argument(
+        "subcommand",
+        type=str,
+        help="Subcommand to execute",
+        choices=("load", "send", "show"),
+    )
+
+    parser.add_argument(
+        "filepath", type=str, help="Path to the file to load", default=None
+    )
+
+    args = parser.parse_args()
+    print(*globals()[args.subcommand](args.filepath))
